refactor: rename encryption module to aes
Rename the encryption module to aes to better reflect its implementation
using AES-128-CBC cipher. Update all module references across the codebase:
- Update lib.rs to use 'mod aes' instead of 'mod encryption'
- Update imports in utils.rs, encrypt.rs, and decrypt.rs
- Keep the same functionality while improving code clarity

This change makes the codebase more maintainable by using more specific
and accurate module naming.
